@@ -1,6 +1,6 @@
 module tailscale.com
 
-go 1.18
+go 1.19
 
 require (
 	filippo.io/mkcert v1.4.3
